backend/requirements.txt
# Core FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database - PostgreSQL for cloud deployment
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# AI & Machine Learning
google-generativeai==0.3.2

# HTTP Client
httpx==0.25.2

# File Processing - Essential only
pillow==10.1.0
moviepy==1.0.3

# Environment & Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# Development & Testing (minimal)
pytest==7.4.3
pytest-asyncio==0.21.1 